kws/requirements.txt
pytorch-lightning==2.0.9
hydra-core==1.2.0
torch==1.13.0
torchaudio==0.13.0
tensorboard
six
pandas
netron
thop